requirements.txt
yfinance==0.2.36
pandas==2.2.0
tabulate==0.9.0
colorama==0.4.6
matplotlib>=3.10.3
streamlit>=1.47.1
pandas-datareader>=0.10.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0